feat: leverage KWayland to handle a lot of Wayland interfaces we shouldn't have to handle
the tldr; is that t wl-client seems to segfault if we don't handle wl_outremoval on the global_remove on the registry. Since we use wlr output manager, we already have _its_ output (head) and spiling on more boilerplate just to make wl-client happy isn't something I want to do.

Fortunately, the KDE folks wrote KWayland, which has its own Registry interface that makes things a bit more C++ / Qtified and it handles a wide range of interfaces. For the ones it doesn't handle like wlr_output_manager, it provides a signal we can connect to in order to get information on it.

So this commit pulls in KWayland to handle all of that for us. May take into use its ConnectionThread if it actually supports QCoreApplication and not just QGuiApplication, but for now the registry is sufficient.

All of this was done just to make it so disabling an output doesn't blow up the daemon.